Improve builder to allow not specifying type in advance.
This makes the DataBuilder class much more pleasant to use
for constructing data recursively, because we do not need
to traverse our recursive structure twice (once to find
the full type and once to add it to the builder), which
would have been necessary previously.

Also adds some tests for DataBuilder to verify that it is
working and to illustrate its usage.