Update manager and server main method to use argot for arg parsing.
This gives a consistent interface for parsing command line arguments
in all top-level binaries. In particular, we add a help option that can
be invoked with -h or --help to get usage information. The usage info
also describes the environment variable fallbacks for the various
command line parameters.

Fixes #26
Review: @strj, @ejeffrey