Add Bytes type ('y') that is distinct from String type ('s')
For now, we allow getting and setting Array[Byte] into 's' field.
This will aid in the transition period until we complete the deployment
of servers and clients that are aware of the new bytes type 'y'.

We also allow Bytes extractor to match both 's' and 'y', and send the
login challenge as 's' like before, until such time as all clients can
handle the new 'y' type.