Block incoming connections until registry is connected.
Most of our labrad code assumes that the registry will be available
when connecting to the manager. In order to support having an external
registry, we block other incoming connections until the registry has.
This is not perfect because of course the registry could disconnect
and reconnect later, but it's better than nothing.