Extend registry context to include a source string.
Like for the multiheaded datavault, when we have a single server that
connects to multiple managers, we need to distinguish contexts created
on each server. To do this, we extend the context key stored internally
to include a srcId string, which is set to the empty string for the
default RegistryActor interface used by the manager's internal registry.