Add support for using OAuth access token instead of id token. (#75)
Previously we used an id token for oauth login, which is a signed blob that includes the user's email address. The manager must call out to an external server to verify this token and extract the email address. The problem with this is that the id token is rather large and usually has a short lifetime.

An access token is a more conventional way to use OAuth; it is a token that the client passes to a server to let the server make API requests on the client's behalf. Here, the manager makes a userinfo requests to get the user's email address and compare it against our db. This is no more work than verifying the id token, but access tokens are smaller and tend to have longer lifetimes, so user's will not need to reauthenticate as often.

Review: @kunalq