Use scala's global ExecutionContext instead of one backed by netty
When using a netty EventLoopGroup we observe the some Futures fail to be
completed properly, leading to timeout errors when Await-ing those
Futures. To avoid this issue, we replace the netty-backed execution
contexts by the standard global execution context, which is similarly
meant for short callbacks that do not block. Blocking server work is
still handled on execution contexts backed by cached thread pools, to
avoid blocking the network or future callback threads.