Allow 0 or 1 parameters to be passed as clusters
Previously, if a setting had 0 parameters it would only accept None, and
if a setting had 1 parameter it would only accept that data. This allows
passing an empty cluster or a cluster with a single element in these two
cases. Eventually, this could allow a model where methods are always
called with a cluster of elements, some of which could be None.