Optimize construction and extraction of array data
We introduce a new `arrayBytes` method on Data objects that returns a
view of the underlying bytes of the array. We then use this to implement
optimized getters and constructors of arrays of primitive fixed-width
types such as int and double. Benchmarks indicate that the getters
(converting from data to `Array[Int]` or `Array[Double]`) are about 10x
faster, while constructors (building Data of type "*i" or "*v" from
scala Arrays) are about 5x faster. In addition, because the `arrayBytes`
method is public, applications can implement their own optimized
routines for other cases, though of course care must be taken to handle
the raw bytes correctly.